Added clearer backlog files
Changed the name of the guidelines file and moved it into documentation folder.

Changed the name of the file with link to temporary backlog and moved it into documentation folder.

Created the go-to backlog file with html formatting. Pulled the user stories from planningpoker and changed them in following way:
Changed the priority of US#23 from 1 to 2, as we don't create the CV screen in priority 1.
Changed the priority of US#49 from 2 to 1, as we will need the characters speed from the beginning.
Removed „melee attack strength modifiers” from US#47-51 as it shouldn't be merged with them.
Added US#81, as it was merged with US#47-51 and should be a separate US; assigned it a point estimate of 21, as all similar US's had the same value.
Removed US#58, 66, 71 and 74 as they were not actual US, and were not estimated.

Pulled the high priority stories from backlog into separate file, so it will be easier to understand.